basic-llm-call/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.0
pytz>=2021.1
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development dependencies
pytest>=6.0.0
black>=21.0.0
isort>=5.0.0
mypy>=0.900